:wrench: try to create a test cluster
