+++ .github/actions/prepare-pulumi/action.yaml
@@ -10,7 +10,7 @@ inputs:
   template:
     required: false
     description: 'source of pulumi template'
-    default: go/library
+    default: phkh-typescript-pulumi-config
 
 runs:
   using: "composite"
@@ -37,12 +37,12 @@ runs:
       uses: actions/setup-node@v3
       with:
         node-version: 20
-      if: startsWith(inputs.template, 'typescript')
+      if: startsWith(inputs.template, 'phkh-typescript')
     - name: Generate nodejs sdk package
       shell: bash
       run: make build_nodejs_sdk
       working-directory: pulumi-component
-      if: startsWith(inputs.template, 'typescript')
+      if: startsWith(inputs.template, 'phkh-typescript')
     - name: Wait for builds
       uses: lewagon/wait-on-check-action@v1.3.1
       with:
@@ -50,7 +50,6 @@ runs:
         check-name: "Build and save component binary"
         repo-token: ${{ github.token }}
         wait-interval: 10
-      if: ${{ inputs.template != 'go/library' }}
     - name: Install pulumi cli
       uses: pulumi/actions@v4
       with:
@@ -61,7 +60,6 @@ runs:
       with:
         workflow: all-build-component.yaml
         workflow_conclusion: success
-      if: ${{ inputs.template != 'go/library' }}
     - name: Add component to pulumi plugin directory
       run: |
         mv -v pulumi-component-*/pulumi-resource-hcloud-kube-hetzner hcloud-kube-hetzner && \
@@ -69,24 +67,28 @@ runs:
         stat hcloud-kube-hetzner && \
         pulumi plugin install resource hcloud-kube-hetzner 0.0.1 -f hcloud-kube-hetzner
       shell: bash
-      if: ${{ inputs.template != 'go/library' }}
     - name: Create a temporary Pulumi project
       run: make test-go-project
       env:
         TEMPLATE: ${{ inputs.template }}
       shell: bash
-      if: startsWith(inputs.template, 'go')
+      if: startsWith(inputs.template, 'phkh-go')
     - name: Create a temporary Pulumi project
       run: make test-ts-project
       env:
         TEMPLATE: ${{ inputs.template }}
       shell: bash
-      if: startsWith(inputs.template, 'typescript')
+      if: startsWith(inputs.template, 'phkh-typescript')
     - name: Downloading GO dependencies 📦
       if: startsWith(inputs.template, 'go')
       shell: bash
       run: go mod download
       working-directory: test-project
+    - name: Include developers tools
+      shell: bash
+      run: |
+        cp -v ../pulumi-templates/dev/go-module-development/pulumi.Makefile ./
+        echo 'include pulumi.Makefile' >> Makefile
     - name: Create a stack
       shell: bash
       if: ${{ inputs.create-stack == 'true'}}
