+++ pulumi-templates/phkh-typescript-pulumi-config/.gitignore
