+++ pulumi-templates/phkh-typescript-pulumi-config/Pulumi.yaml
@@ -1,0 +1,5 @@
+name: ${PROJECT}
+runtime: nodejs
+description: ${DESCRIPTION}
+template:
+  description: "Optimized and Maintenance-free Kubernetes on Hetzner Cloud in one command! with Pulumi!"
