+++ pulumi-templates/phkh-typescript-pulumi-config/cluster-examples/k3s-private-non-ha-ingress.yaml
@@ -1,0 +1,64 @@
+# The cluster has 1 control plane node and 2 agent nodepools.
+# One of the agent nodepool is labeled as ingress and taints the nodes with NoSchedule.
+network:
+  hetzner:
+    enabled: true
+    # It is highly recommended to set a cidr for the private network for at least /20
+    cidr: 10.22.0.0/16
+k8s:
+  kube-api-endpoint:
+    # public|internal
+    type: internal
+nodepools:
+  agents:
+    - pool-id: 01-big-nodes
+      config:
+        server:
+          server-type: cpx31
+      nodes:
+      - node-id: k3s-big-agent01
+    - pool-id: 02-ingress-nodes
+      config:
+        k8s:
+          node-label:
+          - kubernetes.io/role=ingress
+          node-taint:
+            enabled: true
+            taints:
+            - node-role.kubernetes.io/ingress:NoSchedule
+        server:
+          # There is no need to use a big server for ingress nodes
+          # The cpx11 is the minimal for the program
+          server-type: cpx11
+      nodes:
+      - node-id: k3s-ingress01
+  servers:
+    - pool-id: 01-control-plane
+      config:
+        server:
+          server-type: cpx21
+      nodes:
+      - node-id: k3s-server01
+        server:
+          location: fsn1
+defaults:
+  global:
+    k8s:
+      node-label:
+        - k3os.io/upgrade=disabled
+    server:
+      firewall:
+        hetzner:
+          enabled: true
+          allow-icmp: true
+          ssh:
+            allow: true
+    k3s:
+      version: v1.29.4+k3s1
+  servers:
+    k3s:
+      config:
+        disable:
+          - metrics-server
+          - traefik
+          - servicelb
