+++ pulumi-templates/phkh-typescript-pulumi-config/image-builder/microos/template.pkr.hcl
