+++ pulumi-templates/phkh-typescript-pulumi-config/package.json
@@ -5,7 +5,8 @@
         "@types/node": "^16"
     },
     "dependencies": {
-        "@spigell/hcloud-kube-hetzner": "0.0.11",
-        "@pulumi/pulumi": "^3"
+        "@pulumi/pulumi": "^3",
+        "@spigell/hcloud-kube-hetzner": "0.0.12",
+        "yaml": "^2.4.2"
     }
 }
