--- pulumi-templates/typescript/Pulumi.yaml
@@ -1,4 +1,0 @@
-name: ${PROJECT}
-
-runtime: nodejs
-description: Optimized and Maintenance-free Kubernetes on Hetzner Cloud in one command! with Pulumi!
